Fix search/break up into separate files.
