+++ src/main/kotlin/com/openlattice/directory/UserDirectoryService.kt
@@ -23,18 +23,9 @@ package com.openlattice.directory
 
 import com.auth0.json.mgmt.users.User
 import com.codahale.metrics.annotation.Timed
-import com.google.common.collect.ImmutableMap
-import com.hazelcast.core.HazelcastInstance
-import com.hazelcast.core.IMap
-import com.openlattice.auth0.Auth0TokenProvider
-import com.openlattice.authentication.Auth0Configuration
-import com.openlattice.client.RetrofitFactory
-import com.openlattice.datastore.services.Auth0ManagementApi
 import com.openlattice.directory.pojo.Auth0UserBasic
-import com.openlattice.hazelcast.HazelcastMap
-import org.slf4j.LoggerFactory
-import org.springframework.stereotype.Service
 
+internal const val DEFAULT_PAGE_SIZE = 100
 
 interface UserDirectoryService {
     @Timed
@@ -48,73 +39,6 @@ interface UserDirectoryService {
     fun searchAllUsers(searchQuery: String): Map<String, Auth0UserBasic>
 }
 
-private val logger = LoggerFactory.getLogger(UserDirectoryService::class.java)
-private const val DEFAULT_PAGE_SIZE = 100
-
-@Service
-class LocalUserDirectoryService(auth0Configuration: Auth0Configuration) : UserDirectoryService {
-    val users = auth0Configuration.users.associateBy { it.id }
-    override fun getAllUsers(): Map<String, User> {
-        return users
-    }
-
-    override fun getUser(userId: String): User {
-        return users.getValue(userId)
-    }
-
-    override fun searchAllUsers(searchQuery: String): Map<String, Auth0UserBasic> {
-        return users.values.filter { user ->
-            (listOf(user.email, user.name, user.nickname, user.givenName, user.familyName, user.username) +
-                    user.identities.map { it.userId } + user.identities.map { it.connection })
-                    .any { it.contains(searchQuery) }
-        }.map { it.id to Auth0UserBasic(it.id, it.email, it.nickname, it.appMetadata) }.toMap()
-
-    }
-
-}
-
-@Service
-class Auth0UserDirectoryService(
-        auth0TokenProvider: Auth0TokenProvider, hazelcastInstance: HazelcastInstance
-) : UserDirectoryService {
-    private val users = HazelcastMap.USERS.getMap(hazelcastInstance)
-
-    private var auth0ManagementApi = RetrofitFactory
-            .newClient(auth0TokenProvider.managementApiUrl) { auth0TokenProvider.token }
-            .create(Auth0ManagementApi::class.java)
 
 
-    override fun getAllUsers(): Map<String, User> {
-        return ImmutableMap.copyOf(users)
-    }
 
-    override fun getUser(userId: String): User {
-        return users.getValue(userId)
-    }
-
-    //TODO: Switch over to a Hazelcast map to relieve pressure from Auth0
-    override fun searchAllUsers(searchQuery: String): Map<String, Auth0UserBasic> {
-        logger.info("Searching auth0 users with query: $searchQuery")
-
-        var page = 0
-        var pageOfUsers = auth0ManagementApi.searchAllUsers(searchQuery, page++, DEFAULT_PAGE_SIZE)
-        val users = mutableSetOf<Auth0UserBasic>()
-
-        while (pageOfUsers != null) {
-            users.addAll(pageOfUsers)
-
-            if (pageOfUsers.size == DEFAULT_PAGE_SIZE) {
-                pageOfUsers = auth0ManagementApi.searchAllUsers(searchQuery, page++, DEFAULT_PAGE_SIZE)
-            } else {
-                break
-            }
-        }
-
-        if (users.isEmpty()) {
-            logger.warn("Auth0 did not return any users for this search.")
-            return mapOf()
-        }
-
-        return users.map { it.userId to it }.toMap()
-    }
-}
